refactor (client): make edit note page, add new api routes in swagger

+++ client/src/core/api/Notes.ts
@@ -1,5 +1,6 @@
 import { Api } from "../configs/api";
 import { IDataForCreateNote, INote } from "../models/notes";
+import { IUser } from "../models/user";
 
 class NotesApi {
     async getAllNotes(userId: string, { signal }: AbortController) {
@@ -9,8 +10,38 @@ class NotesApi {
         return response;
     }
 
-    async removeNote(noteId: number, { signal }: AbortController) {
-        await Api.delete(`/api/notes/${noteId}`, { signal });
+    async getNote(noteId: number | string, userId: IUser["id"], { signal }) {
+        const response = await Api.get<INote>(
+            `/api/notes/${noteId}?userId=${userId}`,
+            {
+                signal
+            }
+        );
+        return response;
+    }
+
+    async editNote(
+        newNoteData: Omit<INote, "id">,
+        noteId: number | string,
+        userId: IUser["id"],
+        { signal }
+    ) {
+        const response = await Api.put(
+            `/api/note/${noteId}?userId=${userId}`,
+            newNoteData,
+            {
+                signal
+            }
+        );
+        return response;
+    }
+
+    async removeNote(
+        noteId: number,
+        userId: IUser["id"],
+        { signal }: AbortController
+    ) {
+        await Api.delete(`/api/notes/${noteId}?userId=${userId}`, { signal });
     }
 
     async createNote(
